+++ .eslintrc.js
@@ -12,6 +12,8 @@ module.exports = {
     'plugin:ember/recommended'
   ],
   env: {
-    browser: true
+    node: true,
+    browser: true,
+    es6: true
   }
 };