+++ index.js
@@ -1,5 +1,5 @@
 /* eslint-env node */
-'use strict';
+
 var fastbootTransform = require('fastboot-transform');
 
 module.exports = {